third commit
third commit

+++ database/migrations/2019_06_30_002_create_bahan_penolongs_table.php
@@ -14,7 +14,7 @@ class CreateBahanPenolongsTable extends Migration
     public function up()
     {
         Schema::create('bahan_penolong', function (Blueprint $table) {
-            $table->bigIncrements('Id_Bahan_Penolong');
+            $table->increments('Id_Bahan_Penolong');
             $table->string('Nama',50);
             $table->double('Harga');
             $table->timestamps();
